eliminate NAUX_NETWORK in favor of auto-detection instead of hardcoding the number of aux quantities in the network, we now determine it at compile time by parsing the .net file via the new get_naux.py script.  This also separates some of the commonly used network property file stuff into a separate routine.
