--- networks/aprox19/NAUX_NETWORK
@@ -1,5 +1,0 @@
-ifeq ($(USE_NSE),TRUE)
-  NAUX := 3
-else
-  NAUX := 0
-endif
