+++ networks/general_null/Make.package
@@ -29,6 +29,9 @@ else
   fort_opts :=
 endif
 
+NAUX := $(shell $(MICROPHYSICS_HOME)/networks/general_null/get_naux.py --defines "$(DEFINES)" $(GENERAL_NET_INPUTS))
+DEFINES += "-DNAUX_NET=$(NAUX)"
+
 $(NETWORK_OUTPUT_PATH)/network_properties.H: $(GENERAL_NET_INPUTS) $(strip $(NETWORK_TEMPLATE))
 	@echo " "
 	@echo "---------------------------------------------------------------------------"
