--- networks/general_null/NAUX_NETWORK
@@ -1,1 +1,0 @@
-NAUX := 0
