+++ networks/general_null/write_network.py
@@ -3,186 +3,7 @@
 import sys
 import argparse
 
-class Species:
-    """the species class holds the properties of a single species"""
-    def __init__(self):
-        self.name = ""
-        self.short_name = ""
-        self.A = -1
-        self.Z = -1
-        self.is_extra = 0
-
-    def __str__(self):
-        return "species {}, (A,Z) = {},{}".format(self.name, self.A, self.Z)
-
-class AuxVar:
-    """convenience class for an auxilliary variable"""
-    def __init__(self):
-        self.name = ""
-        self.preprocessor = None
-
-    def __str__(self):
-        return "auxillary variable {}".format(self.name)
-
-class UnusedVar:
-    """this is what we return if an Aux var doesn't meet the preprocessor requirements"""
-    def __init__(self):
-        pass
-
-
-def get_next_line(fin):
-    """get_next_line returns the next, non-blank line, with comments
-    stripped"""
-    line = fin.readline()
-
-    pos = str.find(line, "#")
-
-    while (pos == 0 or str.strip(line) == "") and line:
-        line = fin.readline()
-        pos = str.find(line, "#")
-
-    line = line[:pos]
-
-    return line
-
-
-def get_object_index(objs, name):
-    """look through the list and returns the index corresponding to the
-    network object (species or auxvar) specified by name
-
-    """
-
-    index = -1
-
-    for n, o in enumerate(objs):
-        if o.name == name:
-            index = n
-            break
-
-    return index
-
-
-def parse_net_file(species, extra_species, aux_vars, net_file, defines):
-    """parse_net_file read all the species listed in a given network
-    inputs file and adds the valid species to the species list
-
-    """
-
-    err = 0
-
-    try:
-        f = open(net_file, "r")
-    except IOError:
-        sys.exit("write_network.py: ERROR: file "+str(net_file)+" does not exist")
-
-
-    print("write_network.py: working on network file "+str(net_file)+"...")
-
-    line = get_next_line(f)
-
-    while line and not err:
-
-        fields = line.split()
-
-        # read the species or auxiliary variable from the line
-        net_obj, err = parse_network_object(fields, defines)
-        if net_obj is None:
-            return err
-
-        if isinstance(net_obj, UnusedVar):
-            line = get_next_line(f)
-            continue
-
-        objs = species
-        if isinstance(net_obj, AuxVar):
-            objs = aux_vars
-        if isinstance(net_obj, Species):
-            if net_obj.is_extra == 1:
-                objs = extra_species
-
-        # check to see if this species/auxvar is defined in the current list
-        index = get_object_index(objs, net_obj.name)
-
-        if index >= 0:
-            print("write_network.py: ERROR: {} already defined.".format(net_obj))
-            err = 1
-        # add the species or auxvar to the appropriate list
-        objs.append(net_obj)
-
-        line = get_next_line(f)
-
-    return err
-
-
-def parse_network_object(fields, defines):
-    """parse the fields in a line of the network file for either species
-    or auxiliary variables.  Aux variables are prefixed by '__aux_' in
-    the network file. Extra species (that do not participate in the actual
-    RHS, but are used for constructing intermediate rates) are prefix by
-    '__extra_' in the network file.
-
-    """
-
-    err = 0
-
-    # check for aux variables first
-    if fields[0].startswith("__aux_"):
-        ret = AuxVar()
-        ret.name = fields[0][6:]
-        # we can put a preprocessor variable after the aux name to require that it be
-        # set in order to define the auxillary variable
-        try:
-            ret.preprocessor = fields[1]
-        except IndexError:
-            ret.preprocessor = None
-
-        # we can put a preprocessor variable after the aux name to
-        # require that it be set in order to define the auxillary
-        # variable
-        try:
-            ret.preprocessor = fields[1]
-        except IndexError:
-            ret.preprocessor = None
-
-        # if there is a preprocessor attached to this variable, then
-        # we will check if we have defined that
-        if ret.preprocessor is not None:
-            if f"-D{ret.preprocessor }" not in defines:
-                ret = UnusedVar()
-
-        # we can put a preprocessor variable after the aux name to
-        # require that it be set in order to define the auxillary
-        # variable
-        try:
-            ret.preprocessor = fields[1]
-        except IndexError:
-            ret.preprocessor = None
-
-        # if there is a preprocessor attached to this variable, then
-        # we will check if we have defined that
-        if ret.preprocessor is not None:
-            if f"-D{ret.preprocessor }" not in defines:
-                ret = UnusedVar()
-
-    # check for missing fields in species definition
-    elif not len(fields) == 4:
-        print(" ".join(fields))
-        print("write_network.py: ERROR: missing one or more fields in species definition.")
-        ret = None
-        err = 1
-    else:
-        ret = Species()
-
-        if fields[0].startswith("__extra_"):
-            ret.name = fields[0][8:]
-            ret.is_extra = 1
-        else:
-            ret.name = fields[0]
-        ret.short_name = fields[1]
-        ret.A = float(fields[2])
-        ret.Z = float(fields[3])
-
-    return ret, err
+import network_param_file
 
 
 def abort(outfile):
@@ -195,7 +16,6 @@ def abort(outfile):
     fout.write("There was an error parsing the network files")
     fout.close()
     sys.exit(1)
-
 
 
 def write_network(network_template, header_template,
@@ -213,7 +33,9 @@ def write_network(network_template, header_template,
     #-------------------------------------------------------------------------
     # read the species defined in the net_file
     #-------------------------------------------------------------------------
-    err = parse_net_file(species, extra_species, aux_vars, net_file, defines)
+    print(f"write_network.py: working on network file {net_file} ...")
+
+    err = network_param_file.parse(species, extra_species, aux_vars, net_file, defines)
 
     if err:
         abort(network_file)
@@ -243,13 +65,13 @@ def write_network(network_template, header_template,
         if tmp == "":
             continue
 
-        print("writing {}".format(out_file))
+        print(f"writing {out_file}")
 
         # read the template
         try:
-            template = open(tmp, "r")
-        except IOError:
-            sys.exit("write_network.py: ERROR: file {} does not exist".format(tmp))
+            template = open(tmp)
+        except OSError:
+            sys.exit(f"write_network.py: ERROR: file {tmp} does not exist")
         else:
             template_lines = template.readlines()
             template.close()
@@ -279,150 +101,150 @@ def write_network(network_template, header_template,
                 elif keyword == "SPEC_NAMES":
                     if lang == "Fortran":
                         for n, spec in enumerate(species):
-                            fout.write("{}spec_names({}) = \"{}\"\n".format(indent, n+1, spec.name))
+                            fout.write(f"{indent}spec_names({n+1}) = \"{spec.name}\"\n")
 
                     elif lang == "C++":
                         for n, spec in enumerate(species):
-                            fout.write("{}\"{}\",   // {} \n".format(indent, spec.name, n))
+                            fout.write(f"{indent}\"{spec.name}\",   // {n} \n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}\"{}\",   // {} \n".format(indent, spec.name, n + len(species)))
+                            fout.write(f"{indent}\"{spec.name}\",   // {n + len(species)} \n")
 
                 elif keyword == "SHORT_SPEC_NAMES":
                     if lang == "Fortran":
                         for n, spec in enumerate(species):
-                            fout.write("{}short_spec_names({}) = \"{}\"\n".format(indent, n+1, spec.short_name))
+                            fout.write(f"{indent}short_spec_names({n+1}) = \"{spec.short_name}\"\n")
 
                     elif lang == "C++":
                         for n, spec in enumerate(species):
-                            fout.write("{}\"{}\",   // {} \n".format(indent, spec.short_name, n))
+                            fout.write(f"{indent}\"{spec.short_name}\",   // {n} \n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}\"{}\",   // {} \n".format(indent, spec.short_name, n + len(species)))
+                            fout.write(f"{indent}\"{spec.short_name}\",   // {n + len(species)} \n")
 
                 elif keyword == "AION":
                     if lang == "Fortran":
                         for n, spec in enumerate(species):
-                            fout.write("{}aion({}) = {}_rt\n".format(indent, n+1, spec.A))
+                            fout.write(f"{indent}aion({n+1}) = {spec.A}_rt\n")
 
                     elif lang == "C++":
                         for n, spec in enumerate(species):
-                            fout.write("{}{},   // {} \n".format(indent, spec.A, n))
+                            fout.write(f"{indent}{spec.A},   // {n} \n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}{},   // {} \n".format(indent, spec.A, n + len(species)))
+                            fout.write(f"{indent}{spec.A},   // {n + len(species)} \n")
 
                 elif keyword == "AION_CONSTEXPR":
                     if lang == "C++":
                         fout.write("\n")
                         for n, spec in enumerate(species):
-                            fout.write("{}case {}:   // {}\n".format(indent, spec.short_name.capitalize(), n))
-                            fout.write("{}{{\n".format(indent))
-                            fout.write("{}    a = {};\n".format(indent, spec.A))
-                            fout.write("{}    break;\n".format(indent))
-                            fout.write("{}}}\n\n".format(indent))
+                            fout.write(f"{indent}case {spec.short_name.capitalize()}:   // {n}\n")
+                            fout.write(f"{indent}{{\n")
+                            fout.write(f"{indent}    a = {spec.A};\n")
+                            fout.write(f"{indent}    break;\n")
+                            fout.write(f"{indent}}}\n\n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}case {}:   // {}\n".format(indent, spec.short_name.capitalize(), n + len(species)))
-                            fout.write("{}{{\n".format(indent))
-                            fout.write("{}    a = {};\n".format(indent, spec.A))
-                            fout.write("{}    break;\n".format(indent))
-                            fout.write("{}}}\n\n".format(indent))
+                            fout.write(f"{indent}case {spec.short_name.capitalize()}:   // {n + len(species)}\n")
+                            fout.write(f"{indent}{{\n")
+                            fout.write(f"{indent}    a = {spec.A};\n")
+                            fout.write(f"{indent}    break;\n")
+                            fout.write(f"{indent}}}\n\n")
 
                 elif keyword == "AION_INV":
                     if lang == "Fortran":
                         for n, spec in enumerate(species):
-                            fout.write("{}aion_inv({}) = 1.0_rt/{}_rt\n".format(indent, n+1, spec.A))
+                            fout.write(f"{indent}aion_inv({n+1}) = 1.0_rt/{spec.A}_rt\n")
 
                     elif lang == "C++":
                         for n, spec in enumerate(species):
-                            fout.write("{}1.0/{},   // {} \n".format(indent, spec.A, n))
+                            fout.write(f"{indent}1.0/{spec.A},   // {n} \n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}1.0/{},   // {} \n".format(indent, spec.A, n + len(species)))
+                            fout.write(f"{indent}1.0/{spec.A},   // {n + len(species)} \n")
 
                 elif keyword == "ZION":
                     if lang == "Fortran":
                         for n, spec in enumerate(species):
-                            fout.write("{}zion({}) = {}_rt\n".format(indent, n+1, spec.Z))
+                            fout.write(f"{indent}zion({n+1}) = {spec.Z}_rt\n")
 
                     elif lang == "C++":
                         for n, spec in enumerate(species):
-                            fout.write("{}{},   // {}\n".format(indent, spec.Z, n))
+                            fout.write(f"{indent}{spec.Z},   // {n}\n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}{},   // {}\n".format(indent, spec.Z, n + len(species)))
+                            fout.write(f"{indent}{spec.Z},   // {n + len(species)}\n")
 
                 elif keyword == "ZION_CONSTEXPR":
                     if lang == "C++":
                         fout.write("\n")
                         for n, spec in enumerate(species):
-                            fout.write("{}case {}:   // {}\n".format(indent, spec.short_name.capitalize(), n))
-                            fout.write("{}{{\n".format(indent))
-                            fout.write("{}    z = {};\n".format(indent, spec.Z))
-                            fout.write("{}    break;\n".format(indent))
-                            fout.write("{}}}\n\n".format(indent))
+                            fout.write(f"{indent}case {spec.short_name.capitalize()}:   // {n}\n")
+                            fout.write(f"{indent}{{\n")
+                            fout.write(f"{indent}    z = {spec.Z};\n")
+                            fout.write(f"{indent}    break;\n")
+                            fout.write(f"{indent}}}\n\n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}case {}:   // {}\n".format(indent, spec.short_name.capitalize(), n + len(species)))
-                            fout.write("{}{{\n".format(indent))
-                            fout.write("{}    z = {};\n".format(indent, spec.Z))
-                            fout.write("{}    break;\n".format(indent))
-                            fout.write("{}}}\n\n".format(indent))
+                            fout.write(f"{indent}case {spec.short_name.capitalize()}:   // {n + len(species)}\n")
+                            fout.write(f"{indent}{{\n")
+                            fout.write(f"{indent}    z = {spec.Z};\n")
+                            fout.write(f"{indent}    break;\n")
+                            fout.write(f"{indent}}}\n\n")
 
                 elif keyword == "AUX_NAMES":
                     if lang == "Fortran":
                         for n, aux in enumerate(aux_vars):
-                            fout.write("{}aux_names({}) = \"{}\"\n".format(indent, n+1, aux.name))
+                            fout.write(f"{indent}aux_names({n+1}) = \"{aux.name}\"\n")
 
                     elif lang == "C++":
                         for n, aux in enumerate(aux_vars):
-                            fout.write("{}\"{}\",   // {} \n".format(indent, aux.name, n))
+                            fout.write(f"{indent}\"{aux.name}\",   // {n} \n")
 
                 elif keyword == "SHORT_AUX_NAMES":
                     if lang == "Fortran":
                         for n, aux in enumerate(aux_vars):
-                            fout.write("{}short_aux_names({}) = \"{}\"\n".format(indent, n+1, aux.name))
+                            fout.write(f"{indent}short_aux_names({n+1}) = \"{aux.name}\"\n")
 
                     elif lang == "C++":
                         for n, aux in enumerate(aux_vars):
-                            fout.write("{}\"{}\",   // {} \n".format(indent, aux.name, n))
+                            fout.write(f"{indent}\"{aux.name}\",   // {n} \n")
 
                 elif keyword == "PROPERTIES":
                     if lang == "C++":
                         for p in properties:
                             print(p)
-                            fout.write("{}constexpr int {} = {};\n".format(indent, p, properties[p]))
+                            fout.write(f"{indent}constexpr int {p} = {properties[p]};\n")
 
                 elif keyword == "SPECIES_ENUM":
                     if lang == "C++":
                         for n, spec in enumerate(species):
                             if n == 0:
-                                fout.write("{}{}=1,\n".format(indent, spec.short_name.capitalize()))
+                                fout.write(f"{indent}{spec.short_name.capitalize()}=1,\n")
                             else:
-                                fout.write("{}{},\n".format(indent, spec.short_name.capitalize()))
+                                fout.write(f"{indent}{spec.short_name.capitalize()},\n")
                         if len(extra_species) > 0:
-                            fout.write("{}NumberSpecies={},\n".format(indent, species[-1].short_name.capitalize()))
+                            fout.write(f"{indent}NumberSpecies={species[-1].short_name.capitalize()},\n")
                         else:
-                            fout.write("{}NumberSpecies={}\n".format(indent, species[-1].short_name.capitalize()))
+                            fout.write(f"{indent}NumberSpecies={species[-1].short_name.capitalize()}\n")
 
                         for n, spec in enumerate(extra_species):
-                            fout.write("{}{},\n".format(indent, spec.short_name.capitalize()))
+                            fout.write(f"{indent}{spec.short_name.capitalize()},\n")
                         if len(extra_species) > 0:
                             fout.write("{}NumberExtraSpecies={}-{},\n".format(indent,
                                                                               extra_species[-1].short_name.capitalize(),
                                                                               species[-1].short_name.capitalize()))
-                            fout.write("{}NumberTotalSpecies={}\n".format(indent, extra_species[-1].short_name.capitalize()))
+                            fout.write(f"{indent}NumberTotalSpecies={extra_species[-1].short_name.capitalize()}\n")
 
                 elif keyword == "AUXZERO_ENUM":
                     if lang == "C++":
                         if aux_vars:
                             for n, aux in enumerate(aux_vars):
                                 if n == 0:
-                                    fout.write("{}i{}=0,\n".format(indent, aux.name.lower()))
+                                    fout.write(f"{indent}i{aux.name.lower()}=0,\n")
                                 else:
-                                    fout.write("{}i{},\n".format(indent, aux.name.lower()))
-                            fout.write("{}NumberAux=i{}\n".format(indent, aux_vars[-1].name.lower()))
+                                    fout.write(f"{indent}i{aux.name.lower()},\n")
+                            fout.write(f"{indent}NumberAux=i{aux_vars[-1].name.lower()}\n")
 
             else:
                 fout.write(line)
