--- networks/ignition_reaclib/C-burn-simple/NAUX_NETWORK
@@ -1,1 +1,0 @@
-NAUX := 0
