--- networks/ignition_reaclib/URCA-simple/NAUX_NETWORK
@@ -1,1 +1,0 @@
-NAUX := 0
