--- networks/triple_alpha_plus_cago/NAUX_NETWORK
@@ -1,1 +1,0 @@
-NAUX := 0
